--- MomoHanhuaUserData/MomoTranslator_en_US.ts
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<TS version="2.1"><context><name>MainWindow</name><message><location filename="../pyqt5_momoqt_imanhua.py" line="1780"></location><source>Thumbnails</source><translation>Thumbnails</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1785"></location><source>Case Sensitive</source><translation>Case Sensitive</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1791"></location><source>Whole Word</source><translation>Whole Word</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1797"></location><source>Use Regex</source><translation>Use Regex</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1813"></location><source>Search</source><translation>Search</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1824"></location><source>Image List</source><translation>Image List</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1834"></location><source>Layout</source><translation>Layout</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1840"></location><source>Comic</source><translation>Comic</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1841"></location><source>Manga</source><translation>Manga</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1857"></location><source>English</source><translation>English</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1858"></location><source>Chinese Simplified</source><translation>Chinese Simplified</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1859"></location><source>Chinese Traditional</source><translation>Chinese Traditional</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1860"></location><source>Japanese</source><translation>Japanese</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1861"></location><source>Korean</source><translation>Korean</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1868"></location><source>Language</source><translation>Language</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1901"></location><source>Start</source><translation>Start</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1907"></location><source>Analyze Frames</source><translation>Analyze Frames</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1908"></location><source>Analyze Bubbles</source><translation>Analyze Bubbles</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1909"></location><source>OCR</source><translation>OCR</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1910"></location><source>Translate</source><translation>Translate</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1911"></location><source>Lettering</source><translation>Lettering</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1928"></location><source>Setting</source><translation>Setting</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1935"></location><source>Text</source><translation>Text</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1942"></location><source>Layer</source><translation>Layer</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1949"></location><source>Step</source><translation>Step</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1958"></location><source>File</source><translation>File</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2254"></location><location filename="../pyqt5_momoqt_imanhua.py" line="1961"></location><source>Open Folder</source><translation>Open Folder</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2429"></location><location filename="../pyqt5_momoqt_imanhua.py" line="1968"></location><source>Save Image</source><translation>Save Image</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1977"></location><source>Recent Folders</source><translation>Recent Folders</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1981"></location><source>View</source><translation>View</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1990"></location><source>Zoom In</source><translation>Zoom In</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1996"></location><source>Zoom Out</source><translation>Zoom Out</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2005"></location><source>Fit to Screen</source><translation>Fit to Screen</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2013"></location><source>Fit to Width</source><translation>Fit to Width</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2021"></location><source>Fit to Height</source><translation>Fit to Height</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2029"></location><source>Reset Zoom</source><translation>Reset Zoom</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2038"></location><source>Show Thumbnails</source><translation>Show Thumbnails</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2039"></location><source>Show Filenames</source><translation>Show Filenames</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2040"></location><source>Show Both</source><translation>Show Both</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2051"></location><source>Edit</source><translation>Edit</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2054"></location><source>Navigate</source><translation>Navigate</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2057"></location><source>Previous Image</source><translation>Previous Image</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2064"></location><source>Next Image</source><translation>Next Image</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2071"></location><source>First Image</source><translation>First Image</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2078"></location><source>Last Image</source><translation>Last Image</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2095"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2085"></location><source>Help</source><translation>Help</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2544"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2091"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2087"></location><source>About</source><translation>About</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2099"></location><source>Bug Report</source><translation>Bug Report</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2103"></location><source>Update Online</source><translation>Update Online</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2109"></location><source>Toolbar</source><translation>Toolbar</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>Width</source><translation>Width</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>Height</source><translation>Height</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>File Size</source><translation>File Size</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2213"></location><source>Filnename</source><translation>Filnename</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2356"></location><source>Open in Explorer</source><translation>Open in Explorer</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2359"></location><source>Open in Preview</source><translation>Open in Preview</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2366"></location><source>Open with</source><translation>Open with</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2367"></location><source>Photoshop</source><translation>Photoshop</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2371"></location><source>Copy Image Path</source><translation>Copy Image Path</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2372"></location><source>Copy Image Name</source><translation>Copy Image Name</translation></message></context></TS>
