--- MomoHanhuaUserData/MomoTranslator_ja_JP.csv
@@ -1,73 +1,0 @@
-﻿Source,目标语言
-Thumbnails,サムネイル
-Case Sensitive,大文字と小文字を区別する
-Whole Word,完全一致
-Use Regex,正規表現を使用
-Search,検索
-Image List,画像リスト
-Font,フォント
-Font Size,フォントサイズ
-Color,色
-Change Color,色を変更する
-Position,位置
-Transparency,透明度
-Text Items,テキスト項目
-Text,テキスト
-Layer,レイヤー
-Step,ステップ
-File,ファイル
-Open Folder,フォルダを開く
-Save Image,画像を保存
-Recent Folders,最近使用したフォルダ
-View,表示
-Zoom In,ズームイン
-Zoom Out,ズームアウト
-Fit to Screen,画面に合わせる
-Fit to Width,幅に合わせる
-Fit to Height,高さに合わせる
-Reset Zoom,ズームをリセット
-Show Thumbnails,サムネイルを表示
-Show Filenames,ファイル名を表示
-Show Both,両方を表示
-Edit,編集
-Add Text,テキストを追加
-Delete Text,テキストを削除
-Clear All Text,すべてのテキストをクリア
-Navigate,ナビゲート
-Previous Image,前の画像
-Next Image,次の画像
-First Image,最初の画像
-Last Image,最後の画像
-Help,ヘルプ
-Bug Report,バグ報告
-Update Online,オンライン更新
-Toolbar,ツールバー
-Open in Explorer,エクスプローラーで開く
-Open in Preview,プレビューで開く
-Open with,～で開く
-Photoshop,Photoshop
-Copy Image Path,画像パスをコピー
-Copy Image Name,画像名をコピー
-Translation Text,翻訳テキスト
-Text Item,テキスト項目
-Width,幅
-Height,高さ
-File Size,ファイルサイズ
-Filnename,ファイル名
-Start,スタート
-Analyze Frames,マンガのコマを分析する
-Analyze Bubbles,吹き出し解析
-OCR,OCR
-Translate,翻訳
-Lettering,マンガのレタリング
-Setting,設定
-About,について
-Layout,レイアウト
-Comic,アメリカンコミック
-Manga,マンガ
-English,英語
-Chinese Simplified,簡体字中国語
-Chinese Traditional,繁体字中国語
-Japanese,日本語
-Korean,韓国語
-Language,言語
