--- MomoHanhuaUserData/MomoTranslator_ja_JP.ts
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<TS version="2.1"><context><name>MainWindow</name><message><location filename="../pyqt5_momoqt_imanhua.py" line="1780"></location><source>Thumbnails</source><translation>サムネイル</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1785"></location><source>Case Sensitive</source><translation>大文字と小文字を区別する</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1791"></location><source>Whole Word</source><translation>完全一致</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1797"></location><source>Use Regex</source><translation>正規表現を使用</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1813"></location><source>Search</source><translation>検索</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1824"></location><source>Image List</source><translation>画像リスト</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1834"></location><source>Layout</source><translation>レイアウト</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1840"></location><source>Comic</source><translation>アメリカンコミック</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1841"></location><source>Manga</source><translation>マンガ</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1857"></location><source>English</source><translation>英語</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1858"></location><source>Chinese Simplified</source><translation>簡体字中国語</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1859"></location><source>Chinese Traditional</source><translation>繁体字中国語</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1860"></location><source>Japanese</source><translation>日本語</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1861"></location><source>Korean</source><translation>韓国語</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1868"></location><source>Language</source><translation>言語</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1901"></location><source>Start</source><translation>スタート</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1907"></location><source>Analyze Frames</source><translation>マンガのコマを分析する</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1908"></location><source>Analyze Bubbles</source><translation>吹き出し解析</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1909"></location><source>OCR</source><translation>OCR</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1910"></location><source>Translate</source><translation>翻訳</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1911"></location><source>Lettering</source><translation>マンガのレタリング</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1928"></location><source>Setting</source><translation>設定</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1935"></location><source>Text</source><translation>テキスト</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1942"></location><source>Layer</source><translation>レイヤー</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1949"></location><source>Step</source><translation>ステップ</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1958"></location><source>File</source><translation>ファイル</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2254"></location><location filename="../pyqt5_momoqt_imanhua.py" line="1961"></location><source>Open Folder</source><translation>フォルダを開く</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2429"></location><location filename="../pyqt5_momoqt_imanhua.py" line="1968"></location><source>Save Image</source><translation>画像を保存</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1977"></location><source>Recent Folders</source><translation>最近使用したフォルダ</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1981"></location><source>View</source><translation>表示</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1990"></location><source>Zoom In</source><translation>ズームイン</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1996"></location><source>Zoom Out</source><translation>ズームアウト</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2005"></location><source>Fit to Screen</source><translation>画面に合わせる</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2013"></location><source>Fit to Width</source><translation>幅に合わせる</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2021"></location><source>Fit to Height</source><translation>高さに合わせる</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2029"></location><source>Reset Zoom</source><translation>ズームをリセット</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2038"></location><source>Show Thumbnails</source><translation>サムネイルを表示</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2039"></location><source>Show Filenames</source><translation>ファイル名を表示</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2040"></location><source>Show Both</source><translation>両方を表示</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2051"></location><source>Edit</source><translation>編集</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2054"></location><source>Navigate</source><translation>ナビゲート</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2057"></location><source>Previous Image</source><translation>前の画像</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2064"></location><source>Next Image</source><translation>次の画像</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2071"></location><source>First Image</source><translation>最初の画像</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2078"></location><source>Last Image</source><translation>最後の画像</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2095"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2085"></location><source>Help</source><translation>ヘルプ</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2544"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2091"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2087"></location><source>About</source><translation>について</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2099"></location><source>Bug Report</source><translation>バグ報告</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2103"></location><source>Update Online</source><translation>オンライン更新</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2109"></location><source>Toolbar</source><translation>ツールバー</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>Width</source><translation>幅</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>Height</source><translation>高さ</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>File Size</source><translation>ファイルサイズ</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2213"></location><source>Filnename</source><translation>ファイル名</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2356"></location><source>Open in Explorer</source><translation>エクスプローラーで開く</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2359"></location><source>Open in Preview</source><translation>プレビューで開く</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2366"></location><source>Open with</source><translation>～で開く</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2367"></location><source>Photoshop</source><translation>Photoshop</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2371"></location><source>Copy Image Path</source><translation>画像パスをコピー</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2372"></location><source>Copy Image Name</source><translation>画像名をコピー</translation></message></context></TS>
