--- MomoHanhuaUserData/MomoTranslator_ko_KR.csv
@@ -1,73 +1,0 @@
-﻿Source,目标语言
-Thumbnails,썸네일
-Case Sensitive,대소문자 구분
-Whole Word,전체 단어
-Use Regex,정규 표현식 사용
-Search,검색
-Image List,이미지 목록
-Font,글꼴
-Font Size,글꼴 크기
-Color,색상
-Change Color,색상 변경
-Position,위치
-Transparency,투명도
-Text Items,텍스트 항목
-Text,텍스트
-Layer,레이어
-Step,단계
-File,파일
-Open Folder,폴더 열기
-Save Image,이미지 저장
-Recent Folders,최근 폴더
-View,보기
-Zoom In,확대
-Zoom Out,축소
-Fit to Screen,화면에 맞추기
-Fit to Width,너비에 맞추기
-Fit to Height,높이에 맞추기
-Reset Zoom,줌 초기화
-Show Thumbnails,썸네일 표시
-Show Filenames,파일 이름 표시
-Show Both,둘 다 표시
-Edit,편집
-Add Text,텍스트 추가
-Delete Text,텍스트 삭제
-Clear All Text,모든 텍스트 지우기
-Navigate,탐색
-Previous Image,이전 이미지
-Next Image,다음 이미지
-First Image,첫 번째 이미지
-Last Image,마지막 이미지
-Help,도움말
-Bug Report,버그 신고
-Update Online,온라인 업데이트
-Toolbar,툴바
-Open in Explorer,탐색기에서 열기
-Open in Preview,미리보기에서 열기
-Open with,다음으로 열기
-Photoshop,Photoshop
-Copy Image Path,이미지 경로 복사
-Copy Image Name,이미지 이름 복사
-Translation Text,번역 텍스트
-Text Item,텍스트 항목
-Width,너비 
-Height,높이 
-File Size,파일 크기
-Filnename,파일 이름
-Start,시작
-Analyze Frames,만화 프레임 분석
-Analyze Bubbles,버블 분석
-OCR, OCR
-Translate,번역
-Lettering,만화 레터링
-Setting,설정
-About,소개
-Layout,레이아웃
-Comic,미국 만화
-Manga,만화
-English,영어
-Chinese Simplified,중국어 간체자
-Chinese Traditional,중국어 번체자
-Japanese,일본어
-Korean,한국어
-Language,언어
