--- MomoHanhuaUserData/MomoTranslator_ko_KR.ts
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<TS version="2.1"><context><name>MainWindow</name><message><location filename="../pyqt5_momoqt_imanhua.py" line="1780"></location><source>Thumbnails</source><translation>썸네일</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1785"></location><source>Case Sensitive</source><translation>대소문자 구분</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1791"></location><source>Whole Word</source><translation>전체 단어</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1797"></location><source>Use Regex</source><translation>정규 표현식 사용</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1813"></location><source>Search</source><translation>검색</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1824"></location><source>Image List</source><translation>이미지 목록</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1834"></location><source>Layout</source><translation>레이아웃</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1840"></location><source>Comic</source><translation>미국 만화</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1841"></location><source>Manga</source><translation>만화</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1857"></location><source>English</source><translation>영어</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1858"></location><source>Chinese Simplified</source><translation>중국어 간체자</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1859"></location><source>Chinese Traditional</source><translation>중국어 번체자</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1860"></location><source>Japanese</source><translation>일본어</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1861"></location><source>Korean</source><translation>한국어</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1868"></location><source>Language</source><translation>언어</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1901"></location><source>Start</source><translation>시작</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1907"></location><source>Analyze Frames</source><translation>만화 프레임 분석</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1908"></location><source>Analyze Bubbles</source><translation>버블 분석</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1909"></location><source>OCR</source><translation> OCR</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1910"></location><source>Translate</source><translation>번역</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1911"></location><source>Lettering</source><translation>만화 레터링</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1928"></location><source>Setting</source><translation>설정</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1935"></location><source>Text</source><translation>텍스트</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1942"></location><source>Layer</source><translation>레이어</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1949"></location><source>Step</source><translation>단계</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1958"></location><source>File</source><translation>파일</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2254"></location><location filename="../pyqt5_momoqt_imanhua.py" line="1961"></location><source>Open Folder</source><translation>폴더 열기</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2429"></location><location filename="../pyqt5_momoqt_imanhua.py" line="1968"></location><source>Save Image</source><translation>이미지 저장</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1977"></location><source>Recent Folders</source><translation>최근 폴더</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1981"></location><source>View</source><translation>보기</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1990"></location><source>Zoom In</source><translation>확대</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1996"></location><source>Zoom Out</source><translation>축소</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2005"></location><source>Fit to Screen</source><translation>화면에 맞추기</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2013"></location><source>Fit to Width</source><translation>너비에 맞추기</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2021"></location><source>Fit to Height</source><translation>높이에 맞추기</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2029"></location><source>Reset Zoom</source><translation>줌 초기화</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2038"></location><source>Show Thumbnails</source><translation>썸네일 표시</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2039"></location><source>Show Filenames</source><translation>파일 이름 표시</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2040"></location><source>Show Both</source><translation>둘 다 표시</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2051"></location><source>Edit</source><translation>편집</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2054"></location><source>Navigate</source><translation>탐색</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2057"></location><source>Previous Image</source><translation>이전 이미지</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2064"></location><source>Next Image</source><translation>다음 이미지</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2071"></location><source>First Image</source><translation>첫 번째 이미지</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2078"></location><source>Last Image</source><translation>마지막 이미지</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2095"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2085"></location><source>Help</source><translation>도움말</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2544"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2091"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2087"></location><source>About</source><translation>소개</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2099"></location><source>Bug Report</source><translation>버그 신고</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2103"></location><source>Update Online</source><translation>온라인 업데이트</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2109"></location><source>Toolbar</source><translation>툴바</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>Width</source><translation>너비 </translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>Height</source><translation>높이 </translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>File Size</source><translation>파일 크기</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2213"></location><source>Filnename</source><translation>파일 이름</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2356"></location><source>Open in Explorer</source><translation>탐색기에서 열기</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2359"></location><source>Open in Preview</source><translation>미리보기에서 열기</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2366"></location><source>Open with</source><translation>다음으로 열기</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2367"></location><source>Photoshop</source><translation>Photoshop</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2371"></location><source>Copy Image Path</source><translation>이미지 경로 복사</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2372"></location><source>Copy Image Name</source><translation>이미지 이름 복사</translation></message></context></TS>
