--- MomoHanhuaUserData/MomoTranslator_zh_CN.ts
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<TS version="2.1"><context><name>MainWindow</name><message><location filename="../pyqt5_momoqt_imanhua.py" line="1780"></location><source>Thumbnails</source><translation>缩略图</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1785"></location><source>Case Sensitive</source><translation>区分大小写</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1791"></location><source>Whole Word</source><translation>整词匹配</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1797"></location><source>Use Regex</source><translation>使用正则表达式</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1813"></location><source>Search</source><translation>搜索</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1824"></location><source>Image List</source><translation>图像列表</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1834"></location><source>Layout</source><translation>排版</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1840"></location><source>Comic</source><translation>美漫</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1841"></location><source>Manga</source><translation>日漫</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1857"></location><source>English</source><translation>英语</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1858"></location><source>Chinese Simplified</source><translation>简体中文</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1859"></location><source>Chinese Traditional</source><translation>繁体中文</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1860"></location><source>Japanese</source><translation>日语</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1861"></location><source>Korean</source><translation>韩语</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1868"></location><source>Language</source><translation>语言</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1901"></location><source>Start</source><translation>开始</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1907"></location><source>Analyze Frames</source><translation>分析漫画画格</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1908"></location><source>Analyze Bubbles</source><translation>分析气泡</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1909"></location><source>OCR</source><translation>文字识别</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1910"></location><source>Translate</source><translation>翻译</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1911"></location><source>Lettering</source><translation>填字</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1928"></location><source>Setting</source><translation>设置</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1935"></location><source>Text</source><translation>文本</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1942"></location><source>Layer</source><translation>图层</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1949"></location><source>Step</source><translation>步骤</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1958"></location><source>File</source><translation>文件</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2254"></location><location filename="../pyqt5_momoqt_imanhua.py" line="1961"></location><source>Open Folder</source><translation>打开文件夹</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2429"></location><location filename="../pyqt5_momoqt_imanhua.py" line="1968"></location><source>Save Image</source><translation>保存图片</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1977"></location><source>Recent Folders</source><translation>最近文件夹</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1981"></location><source>View</source><translation>查看</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1990"></location><source>Zoom In</source><translation>放大</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1996"></location><source>Zoom Out</source><translation>缩小</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2005"></location><source>Fit to Screen</source><translation>适应屏幕</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2013"></location><source>Fit to Width</source><translation>适应宽度</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2021"></location><source>Fit to Height</source><translation>适应高度</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2029"></location><source>Reset Zoom</source><translation>重置缩放</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2038"></location><source>Show Thumbnails</source><translation>显示缩略图</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2039"></location><source>Show Filenames</source><translation>显示文件名</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2040"></location><source>Show Both</source><translation>全部显示</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2051"></location><source>Edit</source><translation>编辑</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2054"></location><source>Navigate</source><translation>导航</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2057"></location><source>Previous Image</source><translation>上一张图片</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2064"></location><source>Next Image</source><translation>下一张图片</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2071"></location><source>First Image</source><translation>首张图片</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2078"></location><source>Last Image</source><translation>末张图片</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2095"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2085"></location><source>Help</source><translation>帮助</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2544"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2091"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2087"></location><source>About</source><translation>关于</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2099"></location><source>Bug Report</source><translation>报告错误</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2103"></location><source>Update Online</source><translation>在线更新</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2109"></location><source>Toolbar</source><translation>工具栏</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>Width</source><translation>宽度</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>Height</source><translation>高度</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>File Size</source><translation>文件大小</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2213"></location><source>Filnename</source><translation>文件名</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2356"></location><source>Open in Explorer</source><translation>在资源管理器中打开</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2359"></location><source>Open in Preview</source><translation>在预览中打开</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2366"></location><source>Open with</source><translation>打开方式</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2367"></location><source>Photoshop</source><translation>Photoshop</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2371"></location><source>Copy Image Path</source><translation>复制图片路径</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2372"></location><source>Copy Image Name</source><translation>复制图片名称</translation></message></context></TS>
