--- MomoHanhuaUserData/MomoTranslator_zh_TW.ts
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<TS version="2.1"><context><name>MainWindow</name><message><location filename="../pyqt5_momoqt_imanhua.py" line="1780"></location><source>Thumbnails</source><translation>縮圖</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1785"></location><source>Case Sensitive</source><translation>區分大小寫</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1791"></location><source>Whole Word</source><translation>整詞匹配</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1797"></location><source>Use Regex</source><translation>使用正則表達式</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1813"></location><source>Search</source><translation>搜索</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1824"></location><source>Image List</source><translation>圖像列表</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1834"></location><source>Layout</source><translation>排版</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1840"></location><source>Comic</source><translation>美漫</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1841"></location><source>Manga</source><translation>日漫</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1857"></location><source>English</source><translation>英語</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1858"></location><source>Chinese Simplified</source><translation>簡體中文</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1859"></location><source>Chinese Traditional</source><translation>繁體中文</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1860"></location><source>Japanese</source><translation>日語</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1861"></location><source>Korean</source><translation>韓語</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1868"></location><source>Language</source><translation>語言</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1901"></location><source>Start</source><translation>開始</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1907"></location><source>Analyze Frames</source><translation>分析漫畫畫格</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1908"></location><source>Analyze Bubbles</source><translation>分析氣泡</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1909"></location><source>OCR</source><translation>OCR</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1910"></location><source>Translate</source><translation>翻譯</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1911"></location><source>Lettering</source><translation>填字</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1928"></location><source>Setting</source><translation>設置</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1935"></location><source>Text</source><translation>文本</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1942"></location><source>Layer</source><translation>圖層</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1949"></location><source>Step</source><translation>步驟</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1958"></location><source>File</source><translation>文件</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2254"></location><location filename="../pyqt5_momoqt_imanhua.py" line="1961"></location><source>Open Folder</source><translation>打開文件夾</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2429"></location><location filename="../pyqt5_momoqt_imanhua.py" line="1968"></location><source>Save Image</source><translation>保存圖片</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1977"></location><source>Recent Folders</source><translation>最近文件夾</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1981"></location><source>View</source><translation>查看</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1990"></location><source>Zoom In</source><translation>放大</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="1996"></location><source>Zoom Out</source><translation>縮小</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2005"></location><source>Fit to Screen</source><translation>適應屏幕</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2013"></location><source>Fit to Width</source><translation>適應寬度</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2021"></location><source>Fit to Height</source><translation>適應高度</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2029"></location><source>Reset Zoom</source><translation>重置縮放</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2038"></location><source>Show Thumbnails</source><translation>顯示縮圖</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2039"></location><source>Show Filenames</source><translation>顯示文件名</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2040"></location><source>Show Both</source><translation>全部顯示</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2051"></location><source>Edit</source><translation>編輯</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2054"></location><source>Navigate</source><translation>導航</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2057"></location><source>Previous Image</source><translation>上一張圖片</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2064"></location><source>Next Image</source><translation>下一張圖片</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2071"></location><source>First Image</source><translation>第一張圖片</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2078"></location><source>Last Image</source><translation>最後一張圖片</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2095"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2085"></location><source>Help</source><translation>幫助</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2544"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2091"></location><location filename="../pyqt5_momoqt_imanhua.py" line="2087"></location><source>About</source><translation>關於</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2099"></location><source>Bug Report</source><translation>報告錯誤</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2103"></location><source>Update Online</source><translation>在線更新</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2109"></location><source>Toolbar</source><translation>工具欄</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>Width</source><translation>寬度</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>Height</source><translation>高度</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2212"></location><source>File Size</source><translation>檔案大小</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2213"></location><source>Filnename</source><translation>檔案名稱</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2356"></location><source>Open in Explorer</source><translation>在資源管理器中打開</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2359"></location><source>Open in Preview</source><translation>在預覽中打開</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2366"></location><source>Open with</source><translation>用...打開</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2367"></location><source>Photoshop</source><translation>Photoshop</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2371"></location><source>Copy Image Path</source><translation>複製圖片路徑</translation></message><message><location filename="../pyqt5_momoqt_imanhua.py" line="2372"></location><source>Copy Image Name</source><translation>複製圖片名稱</translation></message></context></TS>
