+++ README_20230426.md
@@ -1,0 +1,66 @@
+## 20230426-PR的主要工作
+
+1. 拆分原pyqt_momotranslator.py的3763行代码为以下目录结构，增强可读性与扩展性
+
+```text
+-+- 3rd                                 第三方二进制程序文件夹
+ |
+ +- src -+- core                        核心处理（基础数据结构、图像处理算法）
+ |       | 
+ |       +- ioutils                     IO操作（读写文件）
+ |       |
+ |       +- OCR                         OCR插件（OCR识别应用）
+ |       |
+ |       +- config.py                   全局信息类
+ |       |
+ |       +- gui.py                      程序主UI（pyqt widget）
+ |       |
+ |       +- helper.py                   辅助类
+ |       |
+ |       +- main.py                     主入口
+ |       |
+ |       +- storage.py                  存储类/资源类
+ |       |
+ |       +- utils.py                    工具集
+ |
+ +- tr                                  翻译文件
+```
+
+2. 增加命令行机制，基本使用法为：
+
+    ```bash
+    python main.py -m qt
+    ```
+
+    启动gui前刷新翻译文件：
+
+    ```
+    python main.py -m qt --update_qm
+    ```
+
+    生成req文件：
+
+    ```
+    python main.py -m qt --generate_req
+    ```
+
+3. 增加多语言功能，现可通过菜单项Configuration->Language进行语言切换操作。
+
+4. 持久化GUI设置，位置为DOCUMENT/默墨汉化
+
+5. 其他
+
+    修复PY版本BUG，现已支持Python 3.10。  
+    修复了界面布局问题。  
+    修复了文件识别问题。  
+    修复了PNG格式转四维矩阵导致的算法处理问题。
+
+## 说明
+
+1. 本PR基本保留了原始版本的大部分代码和注释。
+
+2. 本PR去除了原dev模式的代码，那部分代码后续以测试用例的形式加入，新的开发者模式会在算法部分上埋点。
+
+3. 本PR仅在Windows 7/10下的Python 3.9/3.10测试，未对Mac测试，可能需要稍作调整。
+
+4. 处理任务会以多线程进行，优化算法时可能会引入多核优化、CUDA加速，待定。
