+++ requirements.txt
@@ -1,17 +1,16 @@
-baidu_aip
-bs4
 easyocr
-jieba
 loguru
-mammoth
 matplotlib
 natsort
 numpy
-pyqt6
 pytesseract
 qtawesome
 scipy
-stdlib_list
-torch
 webcolors
 xmltodict
+stdlib_list
+PyQt6
+baidu-aip
+chardet
+deep_translator
+python-docx
