+++ src/OCR/OCR.py
@@ -1,0 +1,6 @@
+from abc import ABC, abstractmethod
+
+class OCR(ABC):
+    @abstractmethod
+    def process(self):
+        pass
