+++ src/helper.py
@@ -1,0 +1,33 @@
+from time import time
+from functools import wraps
+from loguru import logger
+
+def timer_decorator(func):
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        start_time = time()
+        result = func(*args, **kwargs)
+        elapsed_time = time() - start_time
+
+        hours, rem = divmod(elapsed_time, 3600)
+        minutes, seconds = divmod(rem, 60)
+
+        if hours > 0:
+            show_run_time = f"{int(hours)}时{int(minutes)}分{seconds:.2f}秒"
+        elif minutes > 0:
+            show_run_time = f"{int(minutes)}分{seconds:.2f}秒"
+        else:
+            show_run_time = f"{seconds:.2f}秒"
+
+        logger.debug(f"{func.__name__} took: {show_run_time}")
+        return result
+
+    return wrapper
+
+class Singleton:
+    _instance = None
+
+    def __new__(cls):
+        if not cls._instance:
+            cls._instance = super().__new__(cls)
+        return cls._instance
