+++ src/ioutils/f.py
@@ -1,0 +1,10 @@
+import os
+
+# ================创建目录================
+
+def make_dir(file_path):
+    if not os.path.exists(file_path):
+        try:
+            os.makedirs(file_path)
+        except BaseException as e:
+            print(e)
