+++ src/ioutils/txt.py
@@ -1,0 +1,47 @@
+from config import GlobalConfig
+
+# ================读取文本================
+
+def read_txt(file_path, encoding='utf-8'):
+    """
+    读取指定文件路径的文本内容。
+
+    :param file_path: 文件路径
+    :param encoding: 文件编码，默认为'utf-8'
+    :return: 返回读取到的文件内容，如果文件不存在则返回None
+    """
+    file_content = None
+    if file_path.exists():
+        with open(file_path, mode='r', encoding=encoding) as file_object:
+            file_content = file_object.read()
+    return file_content
+
+
+# ================写入文件================
+
+def write_txt(file_path, text_input, encoding='utf-8', ignore_empty=True):
+    """
+    将文本内容写入指定的文件路径。
+
+    :param file_path: 文件路径
+    :param text_input: 要写入的文本内容，可以是字符串或字符串列表
+    :param encoding: 文件编码，默认为'utf-8'
+    :param ignore_empty: 是否忽略空内容，默认为True
+    """
+    if text_input:
+        save_text = True
+
+        if isinstance(text_input, list):
+            otext = GlobalConfig.LINE_FEED.join(text_input)
+        else:
+            otext = text_input
+
+        file_content = read_txt(file_path, encoding)
+
+        if file_content == otext or (ignore_empty and otext == ''):
+            save_text = False
+
+        if save_text:
+            with open(file_path, mode='w', encoding=encoding, errors='ignore') as f:
+                f.write(otext)
+
