+++ tr/MomoTranslator_en_US.csv
@@ -1,0 +1,73 @@
+﻿Source,目标语言
+Thumbnails,Thumbnails
+Case Sensitive,Case Sensitive
+Whole Word,Whole Word
+Use Regex,Use Regex
+Search,Search
+Image List,Image List
+Font,Font
+Font Size,Font Size
+Color,Color
+Change Color,Change Color
+Position,Position
+Transparency,Transparency
+Text Items,Text Items
+Text,Text
+Layer,Layer
+Step,Step
+File,File
+Open Folder,Open Folder
+Save Image,Save Image
+Recent Folders,Recent Folders
+View,View
+Zoom In,Zoom In
+Zoom Out,Zoom Out
+Fit to Screen,Fit to Screen
+Fit to Width,Fit to Width
+Fit to Height,Fit to Height
+Reset Zoom,Reset Zoom
+Show Thumbnails,Show Thumbnails
+Show Filenames,Show Filenames
+Show Both,Show Both
+Edit,Edit
+Add Text,Add Text
+Delete Text,Delete Text
+Clear All Text,Clear All Text
+Navigate,Navigate
+Previous Image,Previous Image
+Next Image,Next Image
+First Image,First Image
+Last Image,Last Image
+Help,Help
+Bug Report,Bug Report
+Update Online,Update Online
+Toolbar,Toolbar
+Open in Explorer,Open in Explorer
+Open in Preview,Open in Preview
+Open with,Open with
+Photoshop,Photoshop
+Copy Image Path,Copy Image Path
+Copy Image Name,Copy Image Name
+Translation Text,Translation Text
+Text Item,Text Item
+Width,Width
+Height,Height
+File Size,File Size
+Filnename,Filnename
+Start,Start
+Analyze Frames,Analyze Frames
+Analyze Bubbles,Analyze Bubbles
+OCR,OCR
+Translate,Translate
+Lettering,Lettering
+Setting,Setting
+About,About
+Layout,Layout
+Comic,Comic
+Manga,Manga
+English,English
+Chinese Simplified,Chinese Simplified
+Chinese Traditional,Chinese Traditional
+Japanese,Japanese
+Korean,Korean
+Language,Language
