+++ tr/MomoTranslator_en_US.ts
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="utf-8"?>
+<TS version="2.1"><context><name>GUI</name><message><source>Thumbnails</source><translation>Thumbnails</translation></message><message><source>Case Sensitive</source><translation>Case Sensitive</translation></message><message><source>Whole Word</source><translation>Whole Word</translation></message><message><source>Use Regex</source><translation>Use Regex</translation></message><message><source>Search</source><translation>Search</translation></message><message><source>Image List</source><translation>Image List</translation></message><message><source>Layout</source><translation>Layout</translation></message><message><source>Comic</source><translation>Comic</translation></message><message><source>Manga</source><translation>Manga</translation></message><message><source>English</source><translation>English</translation></message><message><source>Chinese Simplified</source><translation>Chinese Simplified</translation></message><message><source>Chinese Traditional</source><translation>Chinese Traditional</translation></message><message><source>Japanese</source><translation>Japanese</translation></message><message><source>Korean</source><translation>Korean</translation></message><message><source>Language</source><translation>Language</translation></message><message><source>Start</source><translation>Start</translation></message><message><source>Analyze Frames</source><translation>Analyze Frames</translation></message><message><source>Analyze Bubbles</source><translation>Analyze Bubbles</translation></message><message><source>OCR</source><translation>OCR</translation></message><message><source>Translate</source><translation>Translate</translation></message><message><source>Lettering</source><translation>Lettering</translation></message><message><source>Setting</source><translation>Setting</translation></message><message><source>Text</source><translation>Text</translation></message><message><source>Layer</source><translation>Layer</translation></message><message><source>Step</source><translation>Step</translation></message><message><source>File</source><translation>File</translation></message><message><source>Open Folder</source><translation>Open Folder</translation></message><message><source>Save Image</source><translation>Save Image</translation></message><message><source>Recent Folders</source><translation>Recent Folders</translation></message><message><source>View</source><translation>View</translation></message><message><source>Zoom In</source><translation>Zoom In</translation></message><message><source>Zoom Out</source><translation>Zoom Out</translation></message><message><source>Fit to Screen</source><translation>Fit to Screen</translation></message><message><source>Fit to Width</source><translation>Fit to Width</translation></message><message><source>Fit to Height</source><translation>Fit to Height</translation></message><message><source>Reset Zoom</source><translation>Reset Zoom</translation></message><message><source>Show Thumbnails</source><translation>Show Thumbnails</translation></message><message><source>Show Filenames</source><translation>Show Filenames</translation></message><message><source>Show Both</source><translation>Show Both</translation></message><message><source>Edit</source><translation>Edit</translation></message><message><source>Navigate</source><translation>Navigate</translation></message><message><source>Previous Image</source><translation>Previous Image</translation></message><message><source>Next Image</source><translation>Next Image</translation></message><message><source>First Image</source><translation>First Image</translation></message><message><source>Last Image</source><translation>Last Image</translation></message><message><source>Help</source><translation>Help</translation></message><message><source>About</source><translation>About</translation></message><message><source>Bug Report</source><translation>Bug Report</translation></message><message><source>Update Online</source><translation>Update Online</translation></message><message><source>Toolbar</source><translation>Toolbar</translation></message><message><source>Width</source><translation>Width</translation></message><message><source>Height</source><translation>Height</translation></message><message><source>File Size</source><translation>File Size</translation></message><message><source>Filnename</source><translation>Filnename</translation></message><message><source>Open in Explorer</source><translation>Open in Explorer</translation></message><message><source>Open in Preview</source><translation>Open in Preview</translation></message><message><source>Open with</source><translation>Open with</translation></message><message><source>Photoshop</source><translation>Photoshop</translation></message><message><source>Copy Image Path</source><translation>Copy Image Path</translation></message><message><source>Copy Image Name</source><translation>Copy Image Name</translation></message></context></TS>
