+++ tr/MomoTranslator_ja_JP.ts
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="utf-8"?>
+<TS version="2.1"><context><name>GUI</name><message><source>Thumbnails</source><translation>サムネイル</translation></message><message><source>Case Sensitive</source><translation>大文字と小文字を区別する</translation></message><message><source>Whole Word</source><translation>完全一致</translation></message><message><source>Use Regex</source><translation>正規表現を使用</translation></message><message><source>Search</source><translation>検索</translation></message><message><source>Image List</source><translation>画像リスト</translation></message><message><source>Layout</source><translation>レイアウト</translation></message><message><source>Comic</source><translation>アメリカンコミック</translation></message><message><source>Manga</source><translation>マンガ</translation></message><message><source>English</source><translation>英語</translation></message><message><source>Chinese Simplified</source><translation>簡体字中国語</translation></message><message><source>Chinese Traditional</source><translation>繁体字中国語</translation></message><message><source>Japanese</source><translation>日本語</translation></message><message><source>Korean</source><translation>韓国語</translation></message><message><source>Language</source><translation>言語</translation></message><message><source>Start</source><translation>スタート</translation></message><message><source>Analyze Frames</source><translation>マンガのコマを分析する</translation></message><message><source>Analyze Bubbles</source><translation>吹き出し解析</translation></message><message><source>OCR</source><translation>OCR</translation></message><message><source>Translate</source><translation>翻訳</translation></message><message><source>Lettering</source><translation>マンガのレタリング</translation></message><message><source>Setting</source><translation>設定</translation></message><message><source>Text</source><translation>テキスト</translation></message><message><source>Layer</source><translation>レイヤー</translation></message><message><source>Step</source><translation>ステップ</translation></message><message><source>File</source><translation>ファイル</translation></message><message><source>Open Folder</source><translation>フォルダを開く</translation></message><message><source>Save Image</source><translation>画像を保存</translation></message><message><source>Recent Folders</source><translation>最近使用したフォルダ</translation></message><message><source>View</source><translation>表示</translation></message><message><source>Zoom In</source><translation>ズームイン</translation></message><message><source>Zoom Out</source><translation>ズームアウト</translation></message><message><source>Fit to Screen</source><translation>画面に合わせる</translation></message><message><source>Fit to Width</source><translation>幅に合わせる</translation></message><message><source>Fit to Height</source><translation>高さに合わせる</translation></message><message><source>Reset Zoom</source><translation>ズームをリセット</translation></message><message><source>Show Thumbnails</source><translation>サムネイルを表示</translation></message><message><source>Show Filenames</source><translation>ファイル名を表示</translation></message><message><source>Show Both</source><translation>両方を表示</translation></message><message><source>Edit</source><translation>編集</translation></message><message><source>Navigate</source><translation>ナビゲート</translation></message><message><source>Previous Image</source><translation>前の画像</translation></message><message><source>Next Image</source><translation>次の画像</translation></message><message><source>First Image</source><translation>最初の画像</translation></message><message><source>Last Image</source><translation>最後の画像</translation></message><message><source>Help</source><translation>ヘルプ</translation></message><message><source>About</source><translation>について</translation></message><message><source>Bug Report</source><translation>バグ報告</translation></message><message><source>Update Online</source><translation>オンライン更新</translation></message><message><source>Toolbar</source><translation>ツールバー</translation></message><message><source>Width</source><translation>幅</translation></message><message><source>Height</source><translation>高さ</translation></message><message><source>File Size</source><translation>ファイルサイズ</translation></message><message><source>Filnename</source><translation>ファイル名</translation></message><message><source>Open in Explorer</source><translation>エクスプローラーで開く</translation></message><message><source>Open in Preview</source><translation>プレビューで開く</translation></message><message><source>Open with</source><translation>～で開く</translation></message><message><source>Photoshop</source><translation>Photoshop</translation></message><message><source>Copy Image Path</source><translation>画像パスをコピー</translation></message><message><source>Copy Image Name</source><translation>画像名をコピー</translation></message></context></TS>
