+++ tr/MomoTranslator_ko_KR.ts
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="utf-8"?>
+<TS version="2.1"><context><name>GUI</name><message><source>Thumbnails</source><translation>썸네일</translation></message><message><source>Case Sensitive</source><translation>대소문자 구분</translation></message><message><source>Whole Word</source><translation>전체 단어</translation></message><message><source>Use Regex</source><translation>정규 표현식 사용</translation></message><message><source>Search</source><translation>검색</translation></message><message><source>Image List</source><translation>이미지 목록</translation></message><message><source>Layout</source><translation>레이아웃</translation></message><message><source>Comic</source><translation>미국 만화</translation></message><message><source>Manga</source><translation>만화</translation></message><message><source>English</source><translation>영어</translation></message><message><source>Chinese Simplified</source><translation>중국어 간체자</translation></message><message><source>Chinese Traditional</source><translation>중국어 번체자</translation></message><message><source>Japanese</source><translation>일본어</translation></message><message><source>Korean</source><translation>한국어</translation></message><message><source>Language</source><translation>언어</translation></message><message><source>Start</source><translation>시작</translation></message><message><source>Analyze Frames</source><translation>만화 프레임 분석</translation></message><message><source>Analyze Bubbles</source><translation>버블 분석</translation></message><message><source>OCR</source><translation> OCR</translation></message><message><source>Translate</source><translation>번역</translation></message><message><source>Lettering</source><translation>만화 레터링</translation></message><message><source>Setting</source><translation>설정</translation></message><message><source>Text</source><translation>텍스트</translation></message><message><source>Layer</source><translation>레이어</translation></message><message><source>Step</source><translation>단계</translation></message><message><source>File</source><translation>파일</translation></message><message><source>Open Folder</source><translation>폴더 열기</translation></message><message><source>Save Image</source><translation>이미지 저장</translation></message><message><source>Recent Folders</source><translation>최근 폴더</translation></message><message><source>View</source><translation>보기</translation></message><message><source>Zoom In</source><translation>확대</translation></message><message><source>Zoom Out</source><translation>축소</translation></message><message><source>Fit to Screen</source><translation>화면에 맞추기</translation></message><message><source>Fit to Width</source><translation>너비에 맞추기</translation></message><message><source>Fit to Height</source><translation>높이에 맞추기</translation></message><message><source>Reset Zoom</source><translation>줌 초기화</translation></message><message><source>Show Thumbnails</source><translation>썸네일 표시</translation></message><message><source>Show Filenames</source><translation>파일 이름 표시</translation></message><message><source>Show Both</source><translation>둘 다 표시</translation></message><message><source>Edit</source><translation>편집</translation></message><message><source>Navigate</source><translation>탐색</translation></message><message><source>Previous Image</source><translation>이전 이미지</translation></message><message><source>Next Image</source><translation>다음 이미지</translation></message><message><source>First Image</source><translation>첫 번째 이미지</translation></message><message><source>Last Image</source><translation>마지막 이미지</translation></message><message><source>Help</source><translation>도움말</translation></message><message><source>About</source><translation>소개</translation></message><message><source>Bug Report</source><translation>버그 신고</translation></message><message><source>Update Online</source><translation>온라인 업데이트</translation></message><message><source>Toolbar</source><translation>툴바</translation></message><message><source>Width</source><translation>너비 </translation></message><message><source>Height</source><translation>높이 </translation></message><message><source>File Size</source><translation>파일 크기</translation></message><message><source>Filnename</source><translation>파일 이름</translation></message><message><source>Open in Explorer</source><translation>탐색기에서 열기</translation></message><message><source>Open in Preview</source><translation>미리보기에서 열기</translation></message><message><source>Open with</source><translation>다음으로 열기</translation></message><message><source>Photoshop</source><translation>Photoshop</translation></message><message><source>Copy Image Path</source><translation>이미지 경로 복사</translation></message><message><source>Copy Image Name</source><translation>이미지 이름 복사</translation></message></context></TS>
