+++ tr/MomoTranslator_zh_CN.csv
@@ -1,0 +1,76 @@
+﻿Source,目标语言
+Thumbnails,缩略图
+Case Sensitive,区分大小写
+Whole Word,整词匹配
+Use Regex,使用正则表达式
+Search,搜索
+Image List,图像列表
+Font,字体
+Font Size,字体大小
+Color,颜色
+Change Color,更改颜色
+Position,位置
+Transparency,透明度
+Text Items,文本项
+Text,文本
+Layer,图层
+Step,步骤
+File,文件
+Open Folder,打开文件夹
+Save Image,保存图片
+Recent Folders,最近文件夹
+View,查看
+Zoom In,放大
+Zoom Out,缩小
+Fit to Screen,适应屏幕
+Fit to Width,适应宽度
+Fit to Height,适应高度
+Reset Zoom,重置缩放
+Show Thumbnails,显示缩略图
+Show Filenames,显示文件名
+Show Both,全部显示
+Edit,编辑
+Add Text,添加文本
+Delete Text,删除文本
+Clear All Text,清除所有文本
+Navigate,导航
+Previous Image,上一张图片
+Next Image,下一张图片
+First Image,首张图片
+Last Image,末张图片
+Help,帮助
+Bug Report,报告错误
+Update Online,在线更新
+Toolbar,工具栏
+Open in Explorer,在资源管理器中打开
+Open in Preview,在预览中打开
+Open with,打开方式
+Photoshop,Photoshop
+Copy Image Path,复制图片路径
+Copy Image Name,复制图片名称
+Translation Text,翻译文本
+Text Item,文本项
+Width,宽度
+Height,高度
+File Size,文件大小
+Filnename,文件名
+Start,开始
+Analyze Frames,分析漫画画格
+Analyze Bubbles,分析气泡
+OCR,文字识别
+Translate,翻译
+Lettering,填字
+Setting,设置
+About,关于
+Layout,排版
+Comic,美漫
+Manga,日漫
+English,英语
+Chinese Simplified,简体中文
+Chinese Traditional,繁体中文
+Japanese,日语
+Korean,韩语
+Language,语言
+About Qt,
+About iManhua,
+iManhua Help,
