+++ tr/MomoTranslator_zh_CN.ts
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="utf-8"?>
+<TS version="2.1"><context><name>GUI</name><message><source>Thumbnails</source><translation>缩略图</translation></message><message><source>Case Sensitive</source><translation>区分大小写</translation></message><message><source>Whole Word</source><translation>整词匹配</translation></message><message><source>Use Regex</source><translation>使用正则表达式</translation></message><message><source>Search</source><translation>搜索</translation></message><message><source>Image List</source><translation>图像列表</translation></message><message><source>Layout</source><translation>排版</translation></message><message><source>Comic</source><translation>美漫</translation></message><message><source>Manga</source><translation>日漫</translation></message><message><source>English</source><translation>英语</translation></message><message><source>Chinese Simplified</source><translation>简体中文</translation></message><message><source>Chinese Traditional</source><translation>繁体中文</translation></message><message><source>Japanese</source><translation>日语</translation></message><message><source>Korean</source><translation>韩语</translation></message><message><source>Language</source><translation>语言</translation></message><message><source>Start</source><translation>开始</translation></message><message><source>Analyze Frames</source><translation>分析漫画画格</translation></message><message><source>Analyze Bubbles</source><translation>分析气泡</translation></message><message><source>OCR</source><translation>文字识别</translation></message><message><source>Translate</source><translation>翻译</translation></message><message><source>Lettering</source><translation>填字</translation></message><message><source>Setting</source><translation>设置</translation></message><message><source>Text</source><translation>文本</translation></message><message><source>Layer</source><translation>图层</translation></message><message><source>Step</source><translation>步骤</translation></message><message><source>File</source><translation>文件</translation></message><message><source>Open Folder</source><translation>打开文件夹</translation></message><message><source>Save Image</source><translation>保存图片</translation></message><message><source>Recent Folders</source><translation>最近文件夹</translation></message><message><source>View</source><translation>查看</translation></message><message><source>Zoom In</source><translation>放大</translation></message><message><source>Zoom Out</source><translation>缩小</translation></message><message><source>Fit to Screen</source><translation>适应屏幕</translation></message><message><source>Fit to Width</source><translation>适应宽度</translation></message><message><source>Fit to Height</source><translation>适应高度</translation></message><message><source>Reset Zoom</source><translation>重置缩放</translation></message><message><source>Show Thumbnails</source><translation>显示缩略图</translation></message><message><source>Show Filenames</source><translation>显示文件名</translation></message><message><source>Show Both</source><translation>全部显示</translation></message><message><source>Edit</source><translation>编辑</translation></message><message><source>Navigate</source><translation>导航</translation></message><message><source>Previous Image</source><translation>上一张图片</translation></message><message><source>Next Image</source><translation>下一张图片</translation></message><message><source>First Image</source><translation>首张图片</translation></message><message><source>Last Image</source><translation>末张图片</translation></message><message><source>Help</source><translation>帮助</translation></message><message><source>About</source><translation>关于</translation></message><message><source>Bug Report</source><translation>报告错误</translation></message><message><source>Update Online</source><translation>在线更新</translation></message><message><source>Toolbar</source><translation>工具栏</translation></message><message><source>Width</source><translation>宽度</translation></message><message><source>Height</source><translation>高度</translation></message><message><source>File Size</source><translation>文件大小</translation></message><message><source>Filnename</source><translation>文件名</translation></message><message><source>Open in Explorer</source><translation>在资源管理器中打开</translation></message><message><source>Open in Preview</source><translation>在预览中打开</translation></message><message><source>Open with</source><translation>打开方式</translation></message><message><source>Photoshop</source><translation>Photoshop</translation></message><message><source>Copy Image Path</source><translation>复制图片路径</translation></message><message><source>Copy Image Name</source><translation>复制图片名称</translation></message></context></TS>
