+++ tr/MomoTranslator_zh_TW.csv
@@ -1,0 +1,73 @@
+﻿Source,目标语言
+Thumbnails,縮圖
+Case Sensitive,區分大小寫
+Whole Word,整詞匹配
+Use Regex,使用正則表達式
+Search,搜索
+Image List,圖像列表
+Font,字體
+Font Size,字體大小
+Color,顏色
+Change Color,更改顏色
+Position,位置
+Transparency,透明度
+Text Items,文本項
+Text,文本
+Layer,圖層
+Step,步驟
+File,文件
+Open Folder,打開文件夾
+Save Image,保存圖片
+Recent Folders,最近文件夾
+View,查看
+Zoom In,放大
+Zoom Out,縮小
+Fit to Screen,適應屏幕
+Fit to Width,適應寬度
+Fit to Height,適應高度
+Reset Zoom,重置縮放
+Show Thumbnails,顯示縮圖
+Show Filenames,顯示文件名
+Show Both,全部顯示
+Edit,編輯
+Add Text,添加文本
+Delete Text,刪除文本
+Clear All Text,清除所有文本
+Navigate,導航
+Previous Image,上一張圖片
+Next Image,下一張圖片
+First Image,第一張圖片
+Last Image,最後一張圖片
+Help,幫助
+Bug Report,報告錯誤
+Update Online,在線更新
+Toolbar,工具欄
+Open in Explorer,在資源管理器中打開
+Open in Preview,在預覽中打開
+Open with,用...打開
+Photoshop,Photoshop
+Copy Image Path,複製圖片路徑
+Copy Image Name,複製圖片名稱
+Translation Text,翻譯文本
+Text Item,文本項
+Width,寬度
+Height,高度
+File Size,檔案大小
+Filnename,檔案名稱
+Start,開始
+Analyze Frames,分析漫畫畫格
+Analyze Bubbles,分析氣泡
+OCR,OCR
+Translate,翻譯
+Lettering,填字
+Setting,設置
+About,關於
+Layout,排版
+Comic,美漫
+Manga,日漫
+English,英語
+Chinese Simplified,簡體中文
+Chinese Traditional,繁體中文
+Japanese,日語
+Korean,韓語
+Language,語言
