+++ tr/MomoTranslator_zh_TW.ts
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="utf-8"?>
+<TS version="2.1"><context><name>GUI</name><message><source>Thumbnails</source><translation>縮圖</translation></message><message><source>Case Sensitive</source><translation>區分大小寫</translation></message><message><source>Whole Word</source><translation>整詞匹配</translation></message><message><source>Use Regex</source><translation>使用正則表達式</translation></message><message><source>Search</source><translation>搜索</translation></message><message><source>Image List</source><translation>圖像列表</translation></message><message><source>Layout</source><translation>排版</translation></message><message><source>Comic</source><translation>美漫</translation></message><message><source>Manga</source><translation>日漫</translation></message><message><source>English</source><translation>英語</translation></message><message><source>Chinese Simplified</source><translation>簡體中文</translation></message><message><source>Chinese Traditional</source><translation>繁體中文</translation></message><message><source>Japanese</source><translation>日語</translation></message><message><source>Korean</source><translation>韓語</translation></message><message><source>Language</source><translation>語言</translation></message><message><source>Start</source><translation>開始</translation></message><message><source>Analyze Frames</source><translation>分析漫畫畫格</translation></message><message><source>Analyze Bubbles</source><translation>分析氣泡</translation></message><message><source>OCR</source><translation>OCR</translation></message><message><source>Translate</source><translation>翻譯</translation></message><message><source>Lettering</source><translation>填字</translation></message><message><source>Setting</source><translation>設置</translation></message><message><source>Text</source><translation>文本</translation></message><message><source>Layer</source><translation>圖層</translation></message><message><source>Step</source><translation>步驟</translation></message><message><source>File</source><translation>文件</translation></message><message><source>Open Folder</source><translation>打開文件夾</translation></message><message><source>Save Image</source><translation>保存圖片</translation></message><message><source>Recent Folders</source><translation>最近文件夾</translation></message><message><source>View</source><translation>查看</translation></message><message><source>Zoom In</source><translation>放大</translation></message><message><source>Zoom Out</source><translation>縮小</translation></message><message><source>Fit to Screen</source><translation>適應屏幕</translation></message><message><source>Fit to Width</source><translation>適應寬度</translation></message><message><source>Fit to Height</source><translation>適應高度</translation></message><message><source>Reset Zoom</source><translation>重置縮放</translation></message><message><source>Show Thumbnails</source><translation>顯示縮圖</translation></message><message><source>Show Filenames</source><translation>顯示文件名</translation></message><message><source>Show Both</source><translation>全部顯示</translation></message><message><source>Edit</source><translation>編輯</translation></message><message><source>Navigate</source><translation>導航</translation></message><message><source>Previous Image</source><translation>上一張圖片</translation></message><message><source>Next Image</source><translation>下一張圖片</translation></message><message><source>First Image</source><translation>第一張圖片</translation></message><message><source>Last Image</source><translation>最後一張圖片</translation></message><message><source>Help</source><translation>幫助</translation></message><message><source>About</source><translation>關於</translation></message><message><source>Bug Report</source><translation>報告錯誤</translation></message><message><source>Update Online</source><translation>在線更新</translation></message><message><source>Toolbar</source><translation>工具欄</translation></message><message><source>Width</source><translation>寬度</translation></message><message><source>Height</source><translation>高度</translation></message><message><source>File Size</source><translation>檔案大小</translation></message><message><source>Filnename</source><translation>檔案名稱</translation></message><message><source>Open in Explorer</source><translation>在資源管理器中打開</translation></message><message><source>Open in Preview</source><translation>在預覽中打開</translation></message><message><source>Open with</source><translation>用...打開</translation></message><message><source>Photoshop</source><translation>Photoshop</translation></message><message><source>Copy Image Path</source><translation>複製圖片路徑</translation></message><message><source>Copy Image Name</source><translation>複製圖片名稱</translation></message></context></TS>
